chore: fix some explicit-any and TODOs
