+++ web/helpers/atoms/ImportingModel.atom.ts
@@ -1,0 +1,84 @@
+import { ImportingModel } from '@janhq/core'
+import { atom } from 'jotai'
+
+// store the paths of the models that are being imported
+export const importingModelsAtom = atom<ImportingModel[]>([])
+
+export const updateImportingModelProgressAtom = atom(
+  null,
+  (get, set, importId: string, percentage: number) => {
+    const model = get(importingModelsAtom).find((x) => x.importId === importId)
+    if (!model) return
+    const newModel: ImportingModel = {
+      ...model,
+      status: 'IMPORTING',
+      percentage,
+    }
+    const newList = get(importingModelsAtom).map((x) =>
+      x.importId === importId ? newModel : x
+    )
+    set(importingModelsAtom, newList)
+  }
+)
+
+export const setImportingModelErrorAtom = atom(
+  null,
+  (get, set, importId: string, error: string) => {
+    const model = get(importingModelsAtom).find((x) => x.importId === importId)
+    if (!model) return
+    const newModel: ImportingModel = {
+      ...model,
+      status: 'FAILED',
+    }
+
+    console.error(`Importing model ${model} failed`, error)
+    const newList = get(importingModelsAtom).map((m) =>
+      m.importId === importId ? newModel : m
+    )
+    set(importingModelsAtom, newList)
+  }
+)
+
+export const setImportingModelSuccessAtom = atom(
+  null,
+  (get, set, importId: string, modelId: string) => {
+    const model = get(importingModelsAtom).find((x) => x.importId === importId)
+    if (!model) return
+    const newModel: ImportingModel = {
+      ...model,
+      modelId,
+      status: 'IMPORTED',
+      percentage: 1,
+    }
+    const newList = get(importingModelsAtom).map((x) =>
+      x.importId === importId ? newModel : x
+    )
+    set(importingModelsAtom, newList)
+  }
+)
+
+export const updateImportingModelAtom = atom(
+  null,
+  (
+    get,
+    set,
+    importId: string,
+    name: string,
+    description: string,
+    tags: string[]
+  ) => {
+    const model = get(importingModelsAtom).find((x) => x.importId === importId)
+    if (!model) return
+    const newModel: ImportingModel = {
+      ...model,
+      name,
+      importId,
+      description,
+      tags,
+    }
+    const newList = get(importingModelsAtom).map((x) =>
+      x.importId === importId ? newModel : x
+    )
+    set(importingModelsAtom, newList)
+  }
+)
