+++ web/screens/Chat/ModelSetting/index.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import React from 'react'
 
 import SettingComponentBuilder, {
@@ -9,19 +8,17 @@ const ModelSetting = ({
   componentData,
 }: {
   componentData: SettingComponentData[]
-}) => {
-  return (
-    <>
-      {componentData.filter((e) => e.name !== 'prompt_template').length && (
-        <div className="flex flex-col">
-          <SettingComponentBuilder
-            componentData={componentData}
-            selector={(e) => e.name !== 'prompt_template'}
-          />
-        </div>
-      )}
-    </>
-  )
-}
+}) => (
+  <>
+    {componentData.filter((e) => e.name !== 'prompt_template').length && (
+      <div className="flex flex-col">
+        <SettingComponentBuilder
+          componentData={componentData}
+          selector={(e) => e.name !== 'prompt_template'}
+        />
+      </div>
+    )}
+  </>
+)
 
 export default React.memo(ModelSetting)
