+++ web/services/extensionService.ts
@@ -1,5 +1,5 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 'use client'
+
 import { ExtensionTypeEnum } from '@janhq/core'
 
 import { extensionManager } from '@/extension/ExtensionManager'
